Move testgrid go binaries into testgrid/cmd/ subdirectories
This way gazelle is able to manage the Go rules properly.
