+++ testgrid/config/BUILD
@@ -1,12 +1,6 @@
-# gazelle:proto legacy
-# Gazelle gets confused by the config.proto file here, since it is in the same directory
-# but a different package from main.go.
-# Also, we can't even use the default proto rules, since we apply nonstandard tags to the
-# generated code to support YAML.
-
 package(default_visibility = ["//visibility:public"])
 
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 load("@io_bazel_rules_go//proto:def.bzl", "go_proto_library")
 
 proto_library(
@@ -15,8 +9,8 @@ proto_library(
 )
 
 go_proto_library(
-    name = "config_pb",
-    importpath = "k8s.io/test-infra/testgrid/config/config_pb",
+    name = "config_go_proto",
+    importpath = "k8s.io/test-infra/testgrid/config",
     proto = ":config_proto",
 )
 
@@ -26,42 +20,16 @@ genrule(
         "config.yaml",
     ],
     outs = ["testgrid-config"],
-    cmd = "$(location //testgrid/config) --yaml=$< --output=$@",
+    cmd = "$(location //testgrid/cmd/config) --yaml=$< --output=$@",
     tools = [
-        "//testgrid/config",
-    ],
-)
-
-go_binary(
-    name = "config",
-    embed = [":go_default_library"],
-    importpath = "k8s.io/test-infra/testgrid/config",
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["config_test.go"],
-    data = [
-        "config.yaml",
-        "//jenkins/job-configs:all-srcs",
-        "//mungegithub:submit-queue/deployment/kubernetes/configmap.yaml",
-        "//prow:config.yaml",
-    ],
-    embed = [":go_default_library"],
-    importpath = "k8s.io/test-infra/testgrid/config",
-    deps = [
-        ":config_pb",
-        "//prow/config:go_default_library",
-        "//testgrid/config/yaml2proto:go_default_library",
-        "//vendor/github.com/ghodss/yaml:go_default_library",
+        "//testgrid/cmd/config",
     ],
 )
 
 go_library(
     name = "go_default_library",
-    srcs = ["main.go"],
+    embed = [":config_go_proto"],
     importpath = "k8s.io/test-infra/testgrid/config",
-    deps = ["//testgrid/config/yaml2proto:go_default_library"],
 )
 
 filegroup(
@@ -78,4 +46,21 @@ filegroup(
         "//testgrid/config/yaml2proto:all-srcs",
     ],
     tags = ["automanaged"],
+)
+
+go_test(
+    name = "go_default_xtest",
+    srcs = ["config_test.go"],
+    data = [
+        "config.yaml",
+        "//jenkins/job-configs:all-srcs",
+        "//mungegithub:submit-queue/deployment/kubernetes/configmap.yaml",
+        "//prow:config.yaml",
+    ],
+    deps = [
+        ":go_default_library",
+        "//prow/config:go_default_library",
+        "//testgrid/config/yaml2proto:go_default_library",
+        "//vendor/github.com/ghodss/yaml:go_default_library",
+    ],
 )
