+++ testgrid/updater/BUILD
@@ -1,34 +1,24 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 load("@io_bazel_rules_docker//go:image.bzl", "go_image")
 load("@io_bazel_rules_go//proto:def.bzl", "go_proto_library")
 
 proto_library(
     name = "state_proto",
     srcs = ["state.proto"],
+    visibility = ["//visibility:public"],
 )
 
 go_proto_library(
-    name = "state",
+    name = "state_go_proto",
+    importpath = "k8s.io/test-infra/testgrid/updater",
     proto = ":state_proto",
+    visibility = ["//visibility:public"],
 )
 
 go_library(
     name = "go_default_library",
-    srcs = ["main.go"],
+    embed = [":state_go_proto"],
     importpath = "k8s.io/test-infra/testgrid/updater",
-    visibility = ["//visibility:private"],
-    deps = [
-        ":state",
-        "//vendor/cloud.google.com/go/storage:go_default_library",
-    ],
-)
-
-go_binary(
-    name = "updater",
-    embed = [":go_default_library"],
-    importpath = "k8s.io/test-infra/testgrid/updater",
-    pure = "on",
-    tags = ["automanaged"],
     visibility = ["//visibility:public"],
 )
